Implemented possible solution for part 2 of boss attacks.

+++ Car Chase Bullet Hell Game/Model/EntityParameters/MovementParams.cs
@@ -15,7 +15,7 @@ namespace Car_Chase_Bullet_Hell_Game.Model.EntityParameters
         public int? startDegree { get; set; }
         public int? entity { get; set; }
         public double? direction { get; set; }
-        public double? speed { get; set; }
+        public float? speed { get; set; }
         public float? duration { get; set; }
 
         public static MovementParams DeepCopy(MovementParams old)
